(#9) Fix path separator in nuspec file element
Converts the path separator character in the nuspec file source to the
character that is correct for the OS that is running the packing
operation. Changes "\" to "/" on Linux and MacOS systems, and
changes "/" to "\" on Windows systems.

Keeps chocolatey/choco#2166 fixed

Based on this commit:
chocolatey/nuget-chocolatey@979a47d