Set GetManifestFileName to virtual
This allows the it to be overridden to allow for differently named manifest in Chocolatey.
Specifically, when Chocolatey is using side by side, it previously included the version in of
the package in the filename of the manifest, and this allows it to be set to mimic that
previous behavior.