ExtractPackageAsync respect path resolver manifest file name
This sets the remaining two overloads of PackageExtractor.ExtractPackageAsync
to always extract/copy the .nuspec manifest to the filename provided by
the PackagePathResolver.GetManifestFileName method. The first overload
already respected that filename. This allows Chocolatey to override that
method to set the .nuspec filename so as to include the package version
the filename when in side by side mode to keep compatibility with
previous versions of Chocolatey.