(build) Handle case where only BuildTools are installed
There is a scenario where only the BuildTools are installed on a build agent, rather than
a complete installation of Visual Studio. In this scenario, VSWhere by default doesn't
include BuildTools when querying. This is "by design".  As such, if the first attempt at
finding MSBuild return nothing, additionally check specifically for the BuildTools version,
and then gracefully fail if nothing is found.