(build) Switch to new version structure
Move from 6.0.0-zlocal.32767 to something like 3.0.0-alpha-20221121-13.
Here 3.0.0 was chosen since we are moving from version 2.x of
NuGet.Core. The inclusion of the alpha pre-release label matches what
is done on other projects, and the addition of the build counter to the
end of the version number means that we can distinguish between builds
on the same day.  Finally, we moved to _NOT_ use Semantic Version V2
version numbers, as these are not currently supported by Chocolatey.

Running the build can now be done similar to the following:

.\build.ps1 -BuildNumber 13 -BuildDate 20221118 -ReleaseLabel alpha