(#9) Renamed NuGet.Frameworks project
To be Chocolatey.NuGet.Frameworks. Due to a transitive dependency from
one of the Microsoft Unit Testing libraries on the NuGet.Frameworks
NuGet package, it wasn't possible to directly renamed the
NuGet.Frameworks project to Chocolatey.NuGet.Frameworks in the same way
that we have done the other projects.  When we tried this, there was a
conflict between the assemblies, and it wasn't possible to compile.

We attempted to use the standard solutions like introduce an app.config
file to redirect the binding, but none of these approches worked.
Various solutions were put forward, including using an assembly alias
and extern inclusion, as well as completely removing the
NuGet.Frameworks project reference in favour of the upstream
NuGet.Frameworks NuGet package, but all of these were rulred out in
favour of renaming the project and the assemblies, and then updating
the using statements and various other places where the name is used.

It is understood that there will likely be similar changes required
when we pull in upstream code from the NuGet.Client GitHub repository.