(#9) Switch to Chocolatey specific paths for environment folders
This changes the paths returned by NuGet.Environment when the
CHOCOLATEY_VERSION environment variable is set. This variable is set
when Chocolatey sets up it's configuration, so it is always set when
calling NuGet code within Chocolatey. Most of these paths should not
be used, so they are put in a folder under the cache location called
chocolatey-invalid. If this folder exists, then it means that there
is something in NuGet.Client that is not expected. The temp location
is used to get a location to put a lock file, so it is expected to be
used. Usage of this lock file would not interfere with NuGet, so even
if this code was to be bypassed, that part would not conflict with
NuGet.