(#9) Disable telemetry emission
This disables sending of telemetry information to implementations
of the INuGetTelemetryService service, so if a telemetry sending
implementation is added to the NuGet SDK assemblies in the future,
then it will not have anything to send. This disables both sending
of individual telemetry events, and sending a summary as the event
is disposed.