(#508) Fix title element retrieved from oData feed
Within the response that comes back from an oData query, there are two
titles elements.  The first <title> is contained within the main Atom
portion of the response, and the second <d:Title> is contained within
the DataServices namespace.

WIthin the NuGet.Client code base, the first is used as a fall back for
the identity of the package if for some reason the id element does not
exist.  This code has been left in place, however, the title which is
returned to the package information is extracted from the <d:Title>
element, which contains the expected information for a Chocolatey
package.