(#9) Add missing elements to search metadata
When the first work was done to provide Chocolatey specific server
related information from the NuGet.Client libraries, it concentrated on
the following properties:

PackageHash, PackageHashAlgorithm, PackageSize, VersionDownloadCount,
IsApproved, PackageStatus, PackageSubmittedStatus, PackageTestResultStatus,
PackageTestResultStatusDate, PackageValidationResultStatus,
PackageValidationResultDate, PackageCleanupResultDate,
PackageReviewedDate, PackageApprovedDate, PackageReviewer,
IsDownloadCacheAvailable, DownloadCacheDate, DownloadCache

This can be seen in this commit:

https://github.com/NuGet/NuGet.Client/commit/3c891b18af767149dcdc3df65e7cf1e515ba7d08

During testing, it was found that at least the ReleaseNotes property
needed to be included, since this was missing in the output from the
choco info command.  On further conversations, it was decided that a
number of other properties should be added into the NuGet.Client libraries
so that all meaningful information would be included.  These properties
are:

- IsLatestVersion
- IsAbsoluteLatestVersion
- IsPrerelease
- ReleaseNotes
- ProjectSourceUrl
- PackageSourceUrl
- DocsUrl
- MailingListUrl
- BugTrackerUrl
- DownloadCacheStatus
- PackageScanStatus
- PackageScanResultDate
- PackageScanFlagResult