(#9) Change defaults for capabilities of a feed
In the NuGet.Client assemblies, if it isn't possible to ascertain what
methods are supported from a given feed, the default is to assume that
both Search and support fo IsAbsoluteLatestVersion is true.  This is a
change from the original NuGet.Core assembly, where both of these would
default to false if information from the metadata didn't confirm it.

This can be see here:

https://github.com/chocolatey/nuget-chocolatey/blob/a9f040444c705c267363daed59e1bd1e927c2aa9/src/Core/Repositories/DataServiceContextWrapper.cs#L139-L147

And in the usage of these methods:

https://github.com/chocolatey/nuget-chocolatey/blob/a9f040444c705c267363daed59e1bd1e927c2aa9/src/Core/Repositories/DataServicePackageRepository.cs#L168
https://github.com/chocolatey/nuget-chocolatey/blob/a9f040444c705c267363daed59e1bd1e927c2aa9/src/Core/Repositories/DataServicePackageRepository.cs#L229

The decision was taken to perform this override in the catch block,
which should hopefully limit the chances for merge conflicts from the
upstream repository.