(#9) Add sort for DownloadCount when searching
When searching for packages from both Chocolatey CLI and Chocolatey GUI
it is possible to provide a sort order in the form of DownloadCount. This
functionality was missing from the NuGet.Client libraries, so this commit
adds this back in, for V2 feeds.  It is now possible to order by both Id
and DownloadCount, and the correct oData query will be emitted.

Co-authored-by: TheCakeIsNaOH <TheCakeIsNaOH@gmail.com>