(#9) Add methods for Count of packages from search
In addition to being able to list out the packages from a search on a
feed, it is also necessary to know the numbers of packages that are
available, so that correct paging can be applied in applications like
Chocolatey GUI.  This commit add a SearchCountAsync method with an
implementation for for v2 and v3 feeds. It is believed that the v3 feed
will need additional work, but this implementation is aimed at getting
things working in Chocolatey GUI, when using v2 feeds, so additional
work will be done in follow up commits.

Co-authored-by: TheCakeIsNaOH <TheCakeIsNaOH@gmail.com>