(#9) Add check for capabilities to package search
When a package search resource is returned, it should first check to see
if the feed that is being accessed has the capabilities to perform the
search.  If this is not done, saerches can be emitted against a feed
which will fail.  A similar check is performed when using a listResource
so this commit does something very similar for a searchResource.  Within
this commit, a NullLogger is used, which isn't ideal, but without
additional work, and refactoring, this has been deemed as ok for now. A
follow up issue will be created to try to remove the usage of this
NullLogger.

Co-authored-by: TheCakeIsNaOH <TheCakeIsNaOH@gmail.com>