(#9) Add TypeConverter for NuGetVersion class
Within the Chocolatey codebase, there is some explicit handling done of
a serialized instance of the NuGetVersion class into XML. Without a
converter to handle this work, some aspects of the codebase fails to
work correctly.

This commit adds a new NuGetVersionConverter class which is very
similar to the existing SemanticVersionConverter class, and adds an
attribute to the NuGetVersion class so that upstream users know which
class to use when converting instances of this class.