(#9) Add slnf with only projects Chocolatey needs
Chocolatey only uses a portion of the projects that the
NuGet.Client.sln includes.  This commit adds a slnf file with only the
projects, and associated test projects, that Chocolatey is using.  This
makes it much faster to load and build the projects, as we aren't
building projects that we aren't using/editing.