(#9) Set DownloadCache property in metadatabuilder
While testing some functionality of a v2 feed, it was found that the
DownloadCache property wasn't getting set.  Tracing this back through,
it was found that setting of this property wasn't being done in the
same place as other properties.

This commit adds this in, but it hasn't done a sweep to ensure that
other properties haven't also been missed.  This was found during
testing, and simply needs to be fixed now, and later we can do a
complete sweep of all properties.