(#9) Change wording for package resolution warning
In the original warning, which is displayed when a package dependency
is missing for the currently installed packages, it mentions a
packages.config file. This makes sense in the context of NuGet, but not
in the context of Chocolatey. The wording has been changed to make it
clear that the warning related to the packages contained within the
Chocolatey lib folder.