(#9) Add overload for ResolvePackages method
This overload includes an additional parameter for includePrerelease,
to determine whether prerelease packages should be returned or not. For
Chocolatey, prerelease packages should only be included if the --pre
argument is used.

This commit addresses an issue when installing Chocolatey GUI, where
the dotnetfx dependency is resolved to be the prerelease version, when
it shouldn't be.