(#9) New order by for version searching by version
When searching by version number in the 1.x version of Chocolatey, an
additional orderby of "Id, Version desc" is included.  In order to
maintain backwards compatibility, we need to also include this orderby
in the new NuGet.Client based implementation.

This work is similar to the work that was done in this commit, where
the DownloadCount orderby was added:

https://github.com/chocolatey/NuGet.Client/pull/25/commits/d28424ed2d2bf75afe77d775231c50ccc4a78c66

This commit adds the Version property to SearchOrderBy, which can then
be used within chocolatey/choco codebase, when required.