(#9) New order by for version and downloadcount
When searching by both downloadcount (i.e. the order-by-popularity
option) and for a specific version, it makes sense to combine the order
by into something like:

DownloadCount desc, Id, Version desc

rather than simply by version, or by download count.  Ideally, we would
be able to combine the orderby terms based on which enumeration values
are selected, however, then the order of them can't be controlled.

This commit adds a new DownloadCountAndVersion property to the
SearchOrderBy to allow specifically stating that this order by is
required.