(#9) Add new PackageAsync method to ListResource
When fetching information about a single package, in order to find out
the highest available version, NuGet.Client normally has to make
multiple calls via the FindPackagesById method, in order to get all
package versions, and to then determine which was is highest.

To short circuit this, and make only one call, it is possible to use
the Packages() endpoint, providing the exact id to search for, as well
as whether it should include prereleases or not.

This commit adds a new PackageAsync method to the ListResource, which
can then be used to find out the highest available version of a
particular package using one API call.  This can then be consumed
within the Chocolatey/choco codebase, when the
UsePackageRepositoryOptimizations option is set, which makes it behave
in the same way as the 1.x version of Chocolatey.