(#9) Ensure correct order for packages
When using the SearchResource for a local packages feed, there is no
ordering of the packages that are returned.  As a result, when the
CollapseToHighestVersion method (which assumes that the packages are in
Id order) is called, it is possible that the incorrect package version
for the highest available package is returned.  To prevent this, we
need to ensure that when calling the CollapseToHighestVersion method we
first order the packages based on the Identity.Id.

This problem arose when we switched from using the
ListResource.ListAsync method to the SearchResource.SearchAsync method
by default.  The first correctly ordered the packages that are
returned, but the other doesn't, and this manifested itself by some
failing integration tests when running on MacOS.