(#9) Optimize queries when using extra options
In order to make the most optimal oData queries, we need to ensure that
we are only asking for the information that meets our needs.  For
example, when using the command line options for --by-id-only, or
--by-tag-only, or --id-starts-with, we should add this information into
the oData query that is sent out.

This commit adds this into the queries (it will require an additionl
change into the chocolatey/choco codebase to make use of it), and makes
the resulting queries similar to what is sent out in 1.x of choco.exe.