(#9) Update capabilities to check for FindPackagesById
This commit updates the legacy capabilities resource to
store whether the feed has the endpoint FindPackagesById available
or not.

There are some feeds we are aware of that does not include
this endpoint, and we need to be able to support calls to
these feeds. As such we need to store the capabilities
available for the endpoint.