(#9) Change return value from method
In a previous commit, the BuildFieldSearchFilter method was extended to
include additional filter options for things like ByIdOnly, ByTagOnly
and IdStartsWith.  When this method was extened, it was also updated to
only include property clauses if explicitly asked for, since this is
how the function used to work.  In doing this, the return value was
chosen as string.Empty, which felt like the right choice.  However, in
later processing, checks for null were done, and not an empty string,
which then causes problems in how the oData query was constructed.

This commit address that issue by changing the return type from
string.Empty to null.