(build) Attempt to fix failing build
The build is failing both internally and on GitHub Actions.  It works
fine locally, so the suspicion is that a newer version of Visual
Studio, without any older artifacts, is installed on the build agent,
which is causing the issue.  Upgrading to latest VS Framework Version
to see if that addresses the issue.