(build) Only build projects required by Chocolatey
Due to the builds starting to fail both internally and on GitHub
Actions, we need to make some changes to the way the build is done.
The suspicion is that due to a new version of Visual Studio, or perhaps
the SDK, the part of the build the deals with the building of the
Visual Studio Extension is failing.  Since we don't actually need this
project to be built for what Chocolatey needs, let reduce the amount of
projects that are built, being very specific about which projects are
compiled.

This has been achieved by adding a new -ChocolateyBuild parameter,
which is passed into MSBuild as a parameter, and which then chooses
only the projects that we want/need.  This has been tested locally, and
proven to complete, so time to try it out on CI...