(build) Refactor build for upcoming release
In order to get a shippable artifact, we need to build with a release
label of RTM.  To attempt to prevent the need to hard code this value,
attempt to set this based on the name of the branch being built.
Additionally, change the condition for when artifacts are published to
Hermes.  Finally, change the artifact rules to correctly exclude
packages that end with -beta.nupkg, and add additional exclusions for
alpha and rc packages.