(build) Flip logic on conditions
The conditions for a build step work based on all conditions having to
be true.  This isn't how it was originally believed it would be, so
having to switch the logic to exclude the branches that we don't want
to run.  This is far from perfect, and will need to be revised.