(maint) Update shipped API's
Now that we are getting ready to ship a release of NuGet.Client
specifically for usage with Chocolatey, we need to update all the API's
that have been added/changed/removed, etc.  This was done using the
tool in the tool-local folder, by running dotnet run.  This added all
the new entries in teh unshipped to the shipped folder, as well as
created removed files, for the API that have been removed.