(#35) Fix problem with force-dependencies
When running the command choco uninstall vlc --force-dependencies, it
is possible that the command errors out, since the uninstallation of
the package, and all of its dependencies can't be completed, since
there are other packages that share those dependencies.  In 1.x of
Chocolatey CLI, a warning would have been shown saying which
dependencies can't be uninstalled, but the operation would continue
with what packages can be uninstalled.

This commit changes the default logic to throw an exception when there
is a clash, to instead provide a warning, and to remove the conflicting
dependency, so that the overall operation can continue.

Since a change to a public method was required, an overload has been
provided, and the original one marked as obsolete.  This change was
necessary, since a logger was required, in order to emit the warning to
the output.