(#16) Allow usage of SourceCacheContext everywhere
Within the NuGet.Client assemblies, if there is no SourceCachecContext
available, all outgoing queries are immediately subject to an imposed 1
second delay for each query. When running something like choco outdated
with say 100 packages, this means that this operation will take, as a
minimum, over 100 seconds to complete.

We need to be able to accept a SourceCacheContext for each entry point
so that we can control whether caching happens, or not. For the top
level entry point methods, this is a reasonably simply addition of
adding an overloaded method that takes a SourceCachecContext object.
However, there are other places where adding an overload causes a
cascading of updates due to the inheritance structure of the
ResourceProvider class.  However, it has been decided that these changes
are required, since we need to be able to eliminate all imposed 1 second
delays.

This commit add all required overloads to the methods, as well as adding
entries into the unshipped api files, so that these can be tracked. In
addition, within the LegacyFeedCapabilityResourceV2Feed class a specific
cacheKey was provided for the feed capabilities query, since without
this in place, the HTTP Retry logic path was forced to be executed.