(#36) Handle issue with long file paths
When executing certain queries, it is possible to generate a long file
path, that cannot for saved on certain operating systems. While there
are ways at the OS level to enable this support, we should be able to
handle this without this intervention.

This commit makes changes to support attempting the saving of a cache
file, and it this fails, re-attempt it with a hash of the file path,
which is much shorter, and will hopefully work as expected. To allow
for this to work, it was necessary to change the logic to read the file
again by also checking for the presence of the hashed file name, if the
requested file name doesn't exist.