(#40) Handle long file paths when caching files
During testing, it was found that there are cases where a very long
query string was causing a problem when caching the file earlier in the
callstack than the earlier testing.  This problem is caused when a call
into a BCL method throws a long file path exception, even though the
Operating System has been enabled to use long path file support.  Due
to this, we have introduced a check to see if the file path is longer
than the allowed length, and if it is, immediately create the hashed
filename.

This commit adds a new helper method, GetHashedCacheFileName. which is
used to calculate the hashed file name, in a single location, rather
than duplicating the work in multiple places.