(#42) Ensure locks can be created by current user
We need to be able to run queries, and store the results of the query on
disk, in a location that the current user has access to.  When we change
the `cacheLocation` of the Chocolatey CLI configuration, the location of
the lock file used when saving the HTTP query results shouldn't be
impacted.

This changes in this commit create a new folder to store the lock files
that are used when HTTP Cache Queries are persisted to disk.  By taking
control of this, rather than having this location be controlled by the
`cacheLocation` value in the chocolatey.config file, it means that we
have greater control over _where_ these files end up, and more
importantly, can ensure that the current user has permission to create
these files.

As a result of these changes, the following folders will now exist on
disk:

* C:\ProgramData\ChocolateyHttpCache
* C:\ProgramData\ChocolateyHttpCacheLocks
* C:\Users\<name-of-user>\.chocolatey\http-cache
* C:\Users\<name-of-user>\.chocolatey\http-cache-locks

NOTE: At some point, in a later scope of work, we may want to
consolidate the `Chocolatey`, `Chocolatey GUI`, `ChocolateyHttpCache`,
and `ChocolateyHttpCacheLocks`, into one place, rather than having four
different locations, but this is regarded as a larger scope of work than
is required just now.

A new overload for the `ExecuteWithFileLockedAsync` method has been
introduced, which allows for setting of the `BasePath` variable, which
controls where the lock file is created.  This means that we can
directly set this from the HttpSource class, while leaving the other
lock file logic in place.  Within this overload, we ensure to reset the
BasePath back to null, so that the next call into this method will
re-establish what this should be.