(#44) Use parent directory for lock files location
These changes where lock files are created. Instead of having a
single location when creating lock files which may be a location
that can not be written to in some contextes, this commit changes
the behavior to use the same location that the that is modified,
removed or added is expected to be.

This allows us to continue overriding the %TEMP% variable in
Chocolatey CLI, while still allowing users of non-elevated commands
to function as expected, like `choco download` when the cache location
has been set to an admin only location.