(#46) Fix incorrect variable used as lock directory base
A couple of places in the package extractor we incorrectly uses
the target nupkg as the basis of where to store any lock files that
are related to these files.

This causes an issue due to the lock files being created in the same
directory that NuGet.Client removes all files from before downloading
the nupkg file, and causes a unauthorized exception.

Instead, this commit changes it to use the target path that will be set
to the directory instead, which causes the lock files to be created in its
parent directory.