(#46) Fix use of BasePath throwing exception
In some rare cases the use of the BasePath property to store the
original location before overriding it will throw an exception when
this original location is not writable by the current process.

Instead, this commit changes to store the value directly from the
field with a similar name instead to prevent these exceptions from
occuring in these cases.